OH MY GOD IT'S NOT WORKING BUT ITS WORKING
Got AuthorizationView stuff working but won't recognise that the user is logged in when the cookie is added. Must be something to do with the Persisting Authentication script (I did copy and paste it from the default project).

God do I wish that I set it to be authenticated in the first place.