FINALLY WORKING, SOMEWHAT HAVE AN IDEA WHY
Got authentication state working. The client provider was incorrect due to not getting the correct claims from the server in PersistingAuthStateProvider.

Make pages that user cannot access redirect to access-denied page.

Make logout page not accessibly if not logged in.

Cleaned up UserControlller LogIn() function to make it easier to understand.

Added testing and example code to weather page for getting the current client name from the persistingAuthStateProvider on client side