Added Segment Page, BaseLoggedInPage and Gantt Chart Component
Added Segment page which loads SegmentDetails and list of Tasks from TaskController

Added BaseLoggedInPage for all pages that you must be logged in for. Includes AuthView for moving to access-denied page when user isn't logged in and function for getting the UserId from Claims .

Added Gantt Chart that takes in a list of GanttChartObjects that it displays. (Not completed fully)